small riskadvisor server refactoring
