+++ riskadvisor/model.go
@@ -1,0 +1,12 @@
+package riskadvisor
+
+type AdviceRequest struct {
+	Id  int     `json:"id"`
+	Pod api.Pod `json:"pod"`
+}
+
+type AdviceStatus struct {
+	Id     int         `json:"id"`
+	Status string      `json:"status"`
+	Result api.Binding `json:"result"`
+}
